% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMod.R
\name{andrs2}
\alias{andrs2}
\title{Long run variance of errors}
\usage{
andrs2(x)
}
\arguments{
\item{x}{vector of residuals}
}
\value{
a numeric value representing the long run variance of errors
}
\description{
Estimate long run variance of errors by Andrews method
}
\keyword{internal}
