% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{axis_marks}
\alias{axis_marks}
\title{Define the vector of tick marks for the axis of a plot}
\usage{
axis_marks(lx)
}
\arguments{
\item{lx}{integer scalar, representing the length of the x or y vector}
}
\description{
Define the vector of tick marks for the axis of a plot
}
\keyword{internal}
