% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.convergence.clubs.R
\name{dim.convergence.clubs}
\alias{dim.convergence.clubs}
\title{\code{dim} method for S3 object \code{convergence.clubs}}
\usage{
\method{dim}{convergence.clubs}(x, ...)
}
\arguments{
\item{x}{an object of class \code{convergence.clubs}.}

\item{...}{other parameters to pass to function \code{summary()}.}
}
\value{
an integer vector with two values: the first one indicates the number of
clubs, the second one indicates the number of divergent units
}
\description{
\code{dim} method for S3 object \code{convergence.clubs}
}
