% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryGDP.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{GDP of 152 Countries from 1970 to 2003}
\format{A data frame with 152 rows and 35 variables.}
\usage{
data(GDP)
}
\description{
A dataset containing the GDP of 152 Countries over 34 years (Phillips and Sul, 2009).
}
\details{
\describe{
  \item{ID}{Country names (character);}
  \item{Y1970, ..., Y2003}{GDP from year 1970 to 2003.}
}
}
\references{
Phillips, P. C.; Sul, D., 2009. Economic transition and growth. Journal of Applied Econometrics 24 (7), 1153-1185.
}
\keyword{datasets}
