% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeClubs.R
\name{mergeClubs}
\alias{mergeClubs}
\title{Merges a list of clubs created with the function findClubs
by either Phillips and Sul method or von Lyncker and Thoennessen procedure}
\usage{
mergeClubs(clubs, X, IDvar, yearVar, lastT, method = "", divergent = FALSE,
  threshold = -1.65)
}
\arguments{
\item{clubs}{a club list (created by findClub function)}

\item{X}{matrix or dataframe with data}

\item{yearVar}{column indices of year variables}

\item{lastT}{column index of the last time period (clustering variable)}

\item{IDVar}{column index of regions}
}
\description{
Returns as output a list with merged clubs.
}

