% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutsGDP.R
\docType{data}
\name{nutsGDP}
\alias{nutsGDP}
\title{GDP of NUTS regions from 200x to 20xx.......}
\format{A data frame with 268 rows and 16 variables:
\describe{
  \item{NUTS_ID}{ID of NUTS regions (character);}
  \item{Y2000, ..., Y2014}{GDP from year 2000 to 2014 (filtered in order
   to remove business cycles).}
}}
\usage{
nutsGDP
}
\description{
A dataset containing the GDP of 268 NUTS regions over 15 years.
}
\keyword{datasets}
