% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.convergence.clubs.R
\name{print.convergence.clubs}
\alias{print.convergence.clubs}
\title{Print method for S3 object \code{convergence.clubs}}
\usage{
\method{print}{convergence.clubs}(clubs)
}
\description{
Print method for S3 object \code{convergence.clubs}
}
