% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.convergence.clubs.R
\name{print_table}
\alias{print_table}
\title{#'Print method for S3 object \code{summary.convergence.clubs}
#'
#'
#'@param x an object of class \code{summary.convergence.clubs}.
#'@param ... only present for compatibility with function \code{print()}
#'
#'@export}
\usage{
print_table(x, merged)
}
\description{
#'Print method for S3 object \code{summary.convergence.clubs}
#'
#'
#'@param x an object of class \code{summary.convergence.clubs}.
#'@param ... only present for compatibility with function \code{print()}
#'
#'@export
}
\keyword{internal}
