% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_andrews_hac.R
\name{ps_andrews_hac}
\alias{ps_andrews_hac}
\title{Long run variance of errors}
\usage{
ps_andrews_hac(x)
}
\arguments{
\item{x}{vector of residuals}
}
\value{
a numeric value representing the long run variance of errors
}
\description{
Estimate long run variance of errors through Andrews' method
}
\details{
This function computes the long run variance of residuals of an lm model
    by means of Andrews' with Quadratic Spectral kernel and fixed bandwidth.
    The code is an adaptation of Phillips and Sul (2007)'s code,
    which was written in \emph{GAUSS}.
}
\references{
Andrews, D. W., 1991. Heteroskedasticity and autocorrelation consistent covariance matrix estimation.
Econometrica: Journal of the Econometric Society, 817-858.

 Phillips, P. C.; Sul, D., 2007.
Transition modeling and econometric convergence tests. Econometrica 75 (6), 1771-1855.
}
\keyword{internal}
