% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeH.R
\name{computeH}
\alias{computeH}
\title{H values}
\usage{
computeH(X, dataCols, id)
}
\arguments{
\item{X}{matrix or dataframe containing data}

\item{dataCols}{integer vector with the column indices of the data}

\item{id}{optional; row index of regions for which H values are to be computed; 
if missing, all regions are used}
}
\description{
Computes H values (cross sectional variance)
}
