% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryGDP.R
\docType{data}
\name{countryGDP}
\alias{countryGDP}
\title{GDP of NUTS regions from 200x to 20xx.......}
\format{A data frame with 152 rows and 35 variables:}
\usage{
data(countryGDP)
}
\description{
A dataset containing the GDP of 268 NUTS regions over 15 years.
Data were filtered in order  to remove business cycles.
}
\details{
\describe{
  \item{ID}{Country (character);}
  \item{Y1970, ..., Y2003}{GDP from year 1970 to 2003 (filtered in order
   to remove business cycles).}
}
}
\keyword{datasets}
