% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMod_aqsb.R
\name{estimateMod_aqsb}
\alias{estimateMod_aqsb}
\title{Test convergence}
\usage{
estimateMod_aqsb(H, time_trim)
}
\arguments{
\item{H}{vector of H values}

\item{time_trim}{a numeric value between 0 and 1, representing the portion of
time periods to ignore when computing tvalues}
}
\value{
a list containing information about the model used to run the t-test
on the regions in the club: beta coefficient, standard deviation, t-statistics and p-value.
}
\description{
Test convergence
}

