% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClubs.R
\name{findClubs}
\alias{findClubs}
\title{Finds convergence clubs by means of Phillips and Sul  clustering procedure.}
\usage{
findClubs(X, IDvar, yearVar, lastT, cstar = 0)
}
\arguments{
\item{X}{matrix or dataframe containing data}

\item{yearVar}{vector containing the indices of the variable columns (e.g. GDP of years 2000 to 2016)}

\item{lastT}{column index of the last time period, representing the clustering variable}

\item{cstar}{threshold value for test t (c*)}
}
\description{
Finds convergence clubs by means of Phillips and Sul  clustering procedure.
}

