% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDivergent.R
\name{mergeDivergent}
\alias{mergeDivergent}
\title{Merge convergence clubs}
\usage{
mergeDivergent(clubs, X, dataCols, time_trim, threshold = -1.65)
}
\arguments{
\item{clubs}{a club list (created by findClub or mergeClubs functions)}

\item{X}{dataframe containing data}

\item{dataCols}{integer vector with the column indices of the data}

\item{time_trim}{a numeric value between 0 and 1, representing the portion of
time periods to ignore when computing tvalues}
}
\value{
A list of Convergence Clubs, for each club a list is return with the
following objects: \code{id}, a vector containing the row indices
of the regions in the club; \code{model}, a list containing information
about the model used to run the t-test on the regions in the club;
\code{regions}, a vector containing the names of the regions of the club (optional,
only included if it is present in the \code{clubs} object given in input).
}
\description{
Merges a list of clubs created with the function findClubs
by either Phillips and Sul method or von Lyncker and Thoennessen procedure
}
\details{
Returns as output a list with merged clubs.
}
