% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.convergence.clubs.R
\name{plot.convergence.clubs}
\alias{plot.convergence.clubs}
\title{Plot method for S3 x \code{convergence.clubs}}
\usage{
\method{plot}{convergence.clubs}(x, y = NULL, nrows = NULL, ncols = NULL,
  y_fixed = FALSE, ...)
}
\arguments{
\item{x}{an x of class \code{convergence.clubs}.}

\item{y}{unused, added for compatibility with function \code{plot}}

\item{nrows}{number of rows of the graphical layout, if NULL, it is automatically defined}

\item{ncols}{number of columns of the graphical layout, if NULL, it is automatically defined}

\item{y_fixed}{logical, should the scale of the y axis be the same for all plots?}

\item{...}{other parameters to pass to function \code{plot()}.}
}
\description{
Plot method for S3 x \code{convergence.clubs}
}
\details{
\code{nrows} and \code{ncols} are optional parameters used to define the row and
column number for the plot layout. Both or just one of them may be specified.
If none of them is specified, the layout dimension is chosen automatically.
}
\examples{

data("countryGDP")

clubs <- findClubs(countryGDP, dataCols=2:35, regions = 1, refCol=35, time_trim = 1/3,
                   cstar = 0, HACmethod = "AQSB")

plot(clubs)
plot(clubs, y_fixed=TRUE)
plot(clubs, 2,3)
plot(clubs, nrows=3)
plot(clubs, ncols=3)


}
