% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.convergence.clubs.R
\name{print.convergence.clubs}
\alias{print.convergence.clubs}
\title{Print method for S3 object \code{convergence.clubs}}
\usage{
\method{print}{convergence.clubs}(x, ...)
}
\arguments{
\item{x}{an object of class \code{convergence.clubs}.}

\item{...}{other parameters to pass to function \code{summary()}.}
}
\description{
Print method for S3 object \code{convergence.clubs}
}
