% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.convergence.clubs.R
\name{summary.convergence.clubs}
\alias{summary.convergence.clubs}
\title{Summary method for S3 object \code{convergence.clubs}}
\usage{
\method{summary}{convergence.clubs}(clubs)
}
\description{
Summary method for S3 object \code{convergence.clubs}
}
